+++ src/main/java/io/github/astrapi69/gradle/migration/runner/GradleRunConfigurationsCopier.java
@@ -34,7 +34,6 @@ import java.util.Properties;
 import java.util.logging.Logger;
 
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.text.WordUtils;
 
 import io.github.astrapi69.collection.array.ArrayExtensions;
 import io.github.astrapi69.collection.list.ListExtensions;
@@ -53,6 +52,7 @@ import io.github.astrapi69.gradle.migration.extension.DependenciesExtensions;
 import io.github.astrapi69.gradle.migration.info.CopyGradleRunConfigurations;
 import io.github.astrapi69.gradle.migration.info.DependenciesInfo;
 import io.github.astrapi69.io.StreamExtensions;
+import io.github.astrapi69.string.CaseExtensions;
 import io.github.astrapi69.string.StringExtensions;
 
 public class GradleRunConfigurationsCopier
@@ -63,21 +63,18 @@ public class GradleRunConfigurationsCopier
 	public static void copyOnlyRunConfigurations(String sourceProjectName, String targetProjectName,
 		String sourceProjectDirNamePrefix, String targetProjectDirNamePrefix)
 	{
-
-		CopyGradleRunConfigurations copyGradleRunConfigurationsData = GradleRunConfigurationsCopier
-			.newCopyGradleRunConfigurations(sourceProjectName, targetProjectName,
-				sourceProjectDirNamePrefix, targetProjectDirNamePrefix, true, false);
-		GradleRunConfigurationsCopier.of(copyGradleRunConfigurationsData).copy();
+		copyRunConfigurations(sourceProjectName, targetProjectName, sourceProjectDirNamePrefix,
+			targetProjectDirNamePrefix, true, false);
 	}
 
 	public static void copyRunConfigurations(String sourceProjectName, String targetProjectName,
 		String sourceProjectDirNamePrefix, String targetProjectDirNamePrefix,
 		boolean onlyRunConfigurations, boolean runConfigurationsInSameFolder)
 	{
-
 		CopyGradleRunConfigurations copyGradleRunConfigurationsData = GradleRunConfigurationsCopier
 			.newCopyGradleRunConfigurations(sourceProjectName, targetProjectName,
-				sourceProjectDirNamePrefix, targetProjectDirNamePrefix, true, false);
+				sourceProjectDirNamePrefix, targetProjectDirNamePrefix, onlyRunConfigurations,
+				runConfigurationsInSameFolder);
 		GradleRunConfigurationsCopier.of(copyGradleRunConfigurationsData).copy();
 	}
 
@@ -123,15 +120,6 @@ public class GradleRunConfigurationsCopier
 			.sourceFilenamePrefix(sourceFilenamePrefix).targetFilenamePrefix(targetFilenamePrefix)
 			.sourceProjectName(sourceProjectName).targetProjectName(targetProjectName)
 			.runConfigurationsInSameFolder(runConfigurationsInSameFolder).build();
-	}
-
-	@Deprecated
-	public static String getProjectVersionKeyName(String projectName)
-	{
-		String camelCased = WordUtils.capitalizeFully(projectName, new char[] { '-' })
-			.replaceAll("-", "");
-		String projectVersionKeyName = StringExtensions.firstCharacterToLowerCase(camelCased);
-		return projectVersionKeyName + "Version";
 	}
 
 	public static GradleRunConfigurationsCopier of(
@@ -218,8 +206,9 @@ public class GradleRunConfigurationsCopier
 
 	private void externalizeVersionFromBuildGradle(File targetProjectDir) throws IOException
 	{
-		File buildGradle = new File(targetProjectDir, DependenciesInfo.BUILD_GRADLE_NAME);
-		File gradleProperties = new File(targetProjectDir, DependenciesInfo.GRADLE_PROPERTIES_NAME);
+		File buildGradle = new File(targetProjectDir, DependenciesInfo.BUILD_GRADLE_FILENAME);
+		File gradleProperties = new File(targetProjectDir,
+			DependenciesInfo.GRADLE_PROPERTIES_FILENAME);
 		FileFactory.newFile(gradleProperties);
 		String dependenciesContent = DependenciesExtensions.getDependenciesContent(buildGradle);
 		List<String> stringList = getDependenciesAsStringList(dependenciesContent);